added support for Azure Speech

+++ slide2mp4.sh
@@ -1,5 +1,5 @@
 #!/bin/bash
-# Usage: slide2mp4.sh [option] PDF_FILE TXT_FILE LEXICON_FILE OUTPUT_MP4 ["page_num1 page_num2..."]
+# Usage: slide2mp4.sh [option] PDF_FILE TXT_FILE LEXICON_FILE/URL OUTPUT_MP4 ["page_num1 page_num2..."]
 #
 # Copyright (C) 2021 Hirofumi Kojima
 #
@@ -23,6 +23,10 @@ VOICE_ID="Mizuki"
 FONT_NAME="NotoSansCJKjp-Regular"
 FONT_SIZE="14"
 FPS="25"
+# Azure TTS variables
+AZURE_REGION="japaneast"
+AZURE_TTS_VOICE_ID="ja-JP-NanamiNeural"
+AZURE_TTS_SUBS_KEY_FILENAME=~/.azure/tts-subs-keyfile
 
 
 # rm -rf json mp3 mp4 png srt xml
@@ -32,18 +36,22 @@ print_usage ()
 {
 	echo "Description:"
 	echo "	$(basename $0) is a conversion tool, PDF slides to MP4 with audio and subtitles."
-	echo "	$(basename $0) uses Amazon Polly, Text-to-Speech (TTS) service."
-	echo "	$(basename $0) requires the following commands, aws polly, ffmpeg, gm convert, python3, xmllint."
+	echo "	$(basename $0) uses Amazon Polly (default) or Azure Speech, Text-to-Speech (TTS) service."
+	echo "	$(basename $0) requires the following commands, aws polly, aws s3, ffmpeg, ffprobe, gm convert, python3, xmllint."
 	echo "Usage:"
 	echo "	$(basename $0) [option] PDF_FILE TXT_FILE LEXICON_FILE/URL OUTPUT_MP4 ["page_num1 page_num2..."]"
 	echo "Options:"
-	echo "	-h, --help				print this message."
-	echo "	-geo, --geometry			specify the geometry of output mp4 files. (default geometry is \"1280x720\")"
-	echo "	-le, --ffmpeg-loglevel-error		ffmpeg loglevel is \"error\". (default level is \"info\")"
-	echo "	-neural					use Neural format, if possible."
-	echo "	-npc, --no-pdf-convert			don't convert PDF to png."
-	echo "	-ns, --no-subtitles			convert without subtitles."
-	echo "	-vid, --voice-id			specify Amazon Polly voice ID. (default voice ID is \"Mizuki\", Japanese Female)"
+	echo "	-h, --help			print this message."
+	echo "	-geo, --geometry		specify the geometry of output mp4 files. (default geometry is \"1280x720\")"
+	echo "	-le, --ffmpeg-loglevel-error	ffmpeg loglevel is \"error\". (default level is \"info\")"
+	echo "	-neural				use Amazon Polly Neural format, if possible."
+	echo "	-npc, --no-pdf-convert		don't convert PDF to png."
+	echo "	-ns, --no-subtitles		convert without subtitles."
+	echo "	-vid, --voice-id		specify Amazon Polly voice ID. (default voice ID is \"Mizuki\", Japanese Female)"
+	echo "	-azure				use Azure Speech."
+	echo "	-azure-region			specify Azure Region for using Azure Speech. (default Region is \"japaneast\")"
+	echo "	-azure-vid, --azure-voice-id	specify Azure Speech voice name. (default voice name is \"ja-JP-NanamiNeural\")"
+	echo "	-azure-tts-key			specify subscription key file path for Azure Speech. (default file path is \"~/.azure/tts-subs-keyfile\")"
 	echo ""
 	echo "Example1: The following command creates one mp4 file with audio and subtitles, named \"test-output.mp4\"."
 	echo "	$(basename $0) test-slides.pdf test-slides.txt test-lexicon.pls test-output.mp4"
@@ -59,6 +67,12 @@ print_usage ()
 	echo ""
 	echo "Example5: Specify the Neural format, the geometry of output mp4 files (1080p) and Amazon Polly voice ID, Matthew (Male, English, US). Note that the Neural format only works with some voice IDs."
 	echo "	$(basename $0) -geo 1920x1080 -vid Matthew -neural test.pdf test.txt lexicon.pls output.mp4"
+	echo ""
+	echo "Example6: The following command specifies the use of Azure Speech. The subscription key to use Azure Speech must be found in \"~/azure/.tts-subs-keyfile\". When you run this command, \"test-lexicon.pls\" will be temporarily uploaded to Amazon S3."
+	echo "	$(basename $0) -azure test-slides.pdf test-slides.txt test-lexicon.pls test-output.mp4"
+	echo ""
+	echo "Example7: The following command specifies the Azure Region, voice name, subscription keyfile path to use Azure Speech. When using Azure Speech, you can specify public (non-private) URL where you can refer to \"test.pls\"."
+	echo "	$(basename $0) -azure -azure-region centralus -azure-vid en-US-JennyNeural -azure-tts-key test-azure-keyfile test.pdf test.txt http://publicdomain/test.pls output.mp4"
 	exit
 }
 
@@ -68,6 +82,7 @@ RS=$(cat /dev/urandom |base64 |tr -cd "a-z0-9" |fold -w 16 |head -n 1)
 
 
 NS_FLAG=0; NO_CONVERT_FLAG=0; NEURAL_FLAG=0
+AWS_FLAG=1; AZURE_FLAG=0
 FFMPEG_LOG_LEVEL="-loglevel info"
 i=0; arg_num=$#
 while [ $# -gt 0 ]
@@ -86,6 +101,14 @@ do
 		shift; VOICE_ID="$1"; shift
 	elif [ "$1" == "-neural" ]; then
 		NEURAL_FLAG=1; shift
+	elif [ "$1" == "-azure" ]; then
+		AWS_FLAG=0; AZURE_FLAG=1; shift; 
+	elif [ "$1" == "-azure-region" ]; then
+		shift; AZURE_REGION="$1"; shift
+	elif [ "$1" == "-azure-vid" -o "$1" == "--azure-voice-id" ]; then
+		shift; AZURE_TTS_VOICE_ID="$1"; shift
+	elif [ "$1" == "-azure-tts-key" ]; then
+		shift; AZURE_TTS_SUBS_KEY_FILENAME="$1"; shift
 	else
 		i=$(($i+1)); arg[i]="$1"; shift
 	fi
@@ -105,17 +128,28 @@ if [ $arg_num -lt 4 ]; then
 fi
 
 
-if [[ "$LEXICON" =~ https?://* ]]; then
-	LEXICON_FILE=tmp-lexicon-$RS.pls
-	LEXICON_URL="$LEXICON"
-	wget -q "$LEXICON_URL" -O $LEXICON_FILE
-	if [ ! -s $LEXICON_FILE ]; then
-		echo "Lexicon file is empty. Please make sure that the URL to download is corret, $LEXICON_URL."
-		rm -f tmp-lexicon-$RS.pls
-		exit
+LEXICON_URL=""; LEXICON_FILE="$LEXICON"
+if [ $AWS_FLAG -eq 1 ]; then
+
+	if [[ "$LEXICON" =~ https?://* ]]; then
+		LEXICON_FILE=tmp-lexicon-$RS.pls
+		LEXICON_URL="$LEXICON"
+		wget -q "$LEXICON_URL" -O $LEXICON_FILE
+
+		if [ ! -s $LEXICON_FILE ]; then
+			echo "Lexicon file is empty. Please make sure that the URL to download is corret, $LEXICON_URL."
+			rm -f tmp-lexicon-$RS.pls
+			exit
+		fi
+
 	fi
-else
-	LEXICON_FILE="$LEXICON"
+
+elif [ $AZURE_FLAG -eq 1 ]; then
+
+	if [[ "$LEXICON" =~ https?://* ]]; then
+		LEXICON_URL="$LEXICON"
+	fi
+
 fi
 
 
@@ -167,44 +201,137 @@ if [ $NO_CONVERT_FLAG -eq 0 ]; then
 fi
 
 
-LEXICON_NAME=$RS
-aws polly put-lexicon --name $LEXICON_NAME --content file://"$LEXICON_FILE"
-ENGINE=""
-if [ $NEURAL_FLAG -eq 1 ]; then
-	ENGINE="--engine neural"
+if [ $AWS_FLAG -eq 1 ]; then
+
+	LEXICON_NAME=$RS
+	aws polly put-lexicon --name $LEXICON_NAME --content file://"$LEXICON_FILE" 2> tmp-$RS.txt
+
+	if [ -s tmp-$RS.txt ]; then
+		cat tmp-$RS.txt
+		rm -f tmp-$RS.txt tmp-lexicon-$RS.pls
+		exit
+	fi
+
+	ENGINE=""
+	if [ $NEURAL_FLAG -eq 1 ]; then
+		ENGINE="--engine neural"
+	fi
+
+	for i in $PAGES;
+
+	do aws polly synthesize-speech $ENGINE \
+		--lexicon-names $LEXICON_NAME \
+		--text-type ssml \
+		--output-format mp3 \
+		--voice-id $VOICE_ID \
+		--text file://xml/$i.xml \
+		mp3/$i.mp3 1> /dev/null 2> tmp-$RS.txt;
+
+	if [ -s tmp-$RS.txt ]; then
+		echo "There is the following error in executing aws polly, with xml/$i.xml."
+		cat tmp-$RS.txt; rm -f tmp-$RS.txt
+		aws polly delete-lexicon --name $LEXICON_NAME
+		rm -f tmp-lexicon-$RS.pls
+		exit
+	fi
+
+	echo "mp3/$i.mp3 has been created."   
+
+	if [ $NS_FLAG -eq 0 ]; then
+		aws polly synthesize-speech $ENGINE \
+			--lexicon-names $LEXICON_NAME \
+			--text-type ssml \
+			--output-format json \
+			--voice-id $VOICE_ID \
+			--speech-mark-types='["sentence"]' \
+			--text file://xml/$i.xml \
+			json/$i.json &> /dev/null;
+		echo "json/$i.json has been created."
+	fi
+
+	done
+	rm -f tmp-$RS.txt tmp-lexicon-$RS.pls
+	aws polly delete-lexicon --name $LEXICON_NAME
+
+elif [ $AZURE_FLAG -eq 1 ]; then
+
+	AWS_S3_REGION=""; BUCKET_NAME=tmp-lexicon-$RS
+	if [ "$LEXICON_URL" == "" ]; then
+		AWS_S3_REGION="ap-northeast-1"
+		aws s3api create-bucket --bucket "$BUCKET_NAME" --region "$AWS_S3_REGION" --create-bucket-configuration LocationConstraint="$AWS_S3_REGION" 2> tmp-$RS.txt
+
+		if [ -s tmp-$RS.txt ]; then
+			cat tmp-$RS.txt
+			rm -f tmp-$RS.txt transcripts-tmp-$RS.mp4
+			exit
+		fi
+
+		aws s3 cp "$LEXICON_FILE" s3://"$BUCKET_NAME"/ --acl public-read
+		LEXICON_URL="https://"$BUCKET_NAME".s3."$AWS_S3_REGION".amazonaws.com/$LEXICON_FILE"
+		rm -f tmp-$RS.txt
+	fi
+
+	mkdir -p azure-xml; rm -f azure-xml/*
+	mkdir -p azure-txt; rm -f azure-txt/*
+	for i in $PAGES; do "$SLIDE2MP4_DIR"/lib/ssml-aws2azure.sh xml/$i.xml $i $AZURE_TTS_VOICE_ID $LEXICON_URL; done
+
+	mkdir -p azure-mp3; rm -f azure-mp3/*
+	for i in $PAGES;
+	do 
+		NUMS=$(ls azure-xml/$i-*.xml |wc -w |awk '{print $1}')
+		for j in `seq 1 $NUMS`;
+		do
+			"$SLIDE2MP4_DIR"/lib/azure-tts.sh azure-xml/$i-$j.xml "$AZURE_REGION" "$AZURE_TTS_SUBS_KEY_FILENAME" azure-mp3/$i-$j.mp3
+
+			if [ ! -s azure-mp3/$i-$j.mp3 ]; then
+				echo "azure-mp3/$i-$j.mp3 is empty file."
+				echo "Please check azure-xml/$i-$j.xml, Azure $AZURE_REGION Region, $AZURE_TTS_SUBS_KEY_FILENAME."
+				rm -f tmp-lexicon-$RS.pls
+				exit
+			fi
+
+			echo "file azure-mp3/$i-$j.mp3" >> $i-list-$RS.txt
+		done
+	done
+
+	if [ ! $AWS_S3_REGION == "" ]; then
+		aws s3api delete-object --bucket "$BUCKET_NAME" --key "$LEXICON_FILE"
+		aws s3 rb s3://"$BUCKET_NAME"
+	fi
+
+	if [ $NS_FLAG -eq 0 ]; then
+
+		for i in $PAGES;
+		do
+			time_value=0; time_info=0; rm -f json/$i.json
+			NUMS=$(ls azure-xml/$i-*.xml |wc -w |awk '{print $1}')
+
+			for j in `seq 1 $NUMS`;
+			do
+				seconds=$(ffprobe -loglevel error -hide_banner -show_entries format=duration azure-mp3/$i-$j.mp3 |grep -i duration |sed -e 's/duration=//')
+				mseconds=`echo "scale=4; $seconds * 1000" |bc`
+				python3 "$SLIDE2MP4_DIR"/lib/repr.py azure-txt/$i-$j.txt tmp-$RS.txt
+				awk '{print substr($0, 2, length($0)-2)}' tmp-$RS.txt > value-$RS.txt
+				awk '{print "{\"time\":'$time_info',\"value\":\"" $0}' value-$RS.txt |sed '$s/$/\"}/' >> json/$i.json
+				time_value=`echo "scale=4; $time_value + $mseconds" |bc`
+				time_info=${time_value%.*}
+			done
+
+			echo "json/$i.json has been created."
+
+		done
+
+	fi
+	rm -f tmp-$RS.txt value-$RS.txt
+
+	for i in $PAGES
+	do
+		ffmpeg $FFMPEG_LOG_LEVEL -y -f concat -i $i-list-$RS.txt -c copy mp3/$i.mp3
+		echo "mp3/$i.mp3 has been created."
+	done
+	rm -rf *-list-$RS.txt tmp-lexicon-$RS.pls azure-mp3 azure-txt azure-xml
+
 fi
-for i in $PAGES;
-do aws polly synthesize-speech $ENGINE \
-       --lexicon-names $LEXICON_NAME \
-       --text-type ssml \
-       --output-format mp3 \
-       --voice-id $VOICE_ID \
-       --text file://xml/$i.xml \
-       mp3/$i.mp3 1> /dev/null 2> tmp-$RS.txt;
-   
-   if [ -s tmp-$RS.txt ]; then
-        echo "There is the following error in executing aws polly, with xml/$i.xml."
-        cat tmp-$RS.txt; rm -f tmp-$RS.txt
-        aws polly delete-lexicon --name $LEXICON_NAME
-        exit
-   fi
-
-   echo "mp3/$i.mp3 has been created."   
-
-   if [ $NS_FLAG -eq 0 ]; then
-   	aws polly synthesize-speech $ENGINE \
-            --lexicon-names $LEXICON_NAME \
-            --text-type ssml \
-            --output-format json \
-            --voice-id $VOICE_ID \
-            --speech-mark-types='["sentence"]' \
-            --text file://xml/$i.xml \
-            json/$i.json &> /dev/null;
-	echo "json/$i.json has been created."
-   fi
-done
-rm -f tmp-$RS.txt tmp-lexicon-$RS.pls
-aws polly delete-lexicon --name $LEXICON_NAME
 
 
 if [ $NS_FLAG -eq 0 ]; then
